Additional cleanup of `mudlog`
